+++ Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 actix-web = "3"
+actix-rt = "1.1.1"
 actix-files = "0.3"
 env_logger = "0.7"
 actix-http = "2"
